lock index/findings files while writing ram
